+++ src/model.ml
@@ -222,7 +222,8 @@ type ('id, 'term) mterm_node  =
   | Mtransfer         of 'term transfer_kind_gen
   (* entrypoint *)
   | Mentrypoint       of type_ * 'id * 'term * 'term option  (* type * address * string * require fail *)
-  | Mself             of 'id                   (* entryname *)
+  | Mcallview         of type_ * 'term * 'id * 'term         (* type * string * address * argument *)
+  | Mself             of 'id                                 (* entryname *)
   (* operation *)
   | Moperations
   | Mmkoperation      of 'term * 'term * 'term  (* value * address * args *)
@@ -1374,6 +1375,7 @@ let cmp_mterm_node
     | Mtransfer tr1, Mtransfer tr2                                                     -> cmp_transfer_kind tr1 tr2
     (* entrypoint *)
     | Mentrypoint (t1, a1, s1, r1), Mentrypoint (t2, a2, s2, r2)                       -> cmp_type t1 t2 && cmpi a1 a2 && cmp s1 s2 && Option.cmp cmp r1 r2
+    | Mcallview (t1, a1, b1, c1), Mcallview (t2, a2, b2, c2)                           -> cmp_type t1 t2 && cmp a1 a2 && cmpi b1 b2 && cmp c1 c2
     | Mself id1, Mself id2                                                             -> cmpi id1 id2
     (* operation *)
     | Moperations, Moperations                                                         -> true
@@ -1798,6 +1800,7 @@ let map_term_node_internal (fi : ident -> ident) (g : 'id -> 'id) (ft : type_ ->
   | Mtransfer tr                   -> Mtransfer (map_transfer_kind fi ft f tr)
   (* entrypoint *)
   | Mentrypoint (t, a, s, r)       -> Mentrypoint (ft t, g a, f s, Option.map f r)
+  | Mcallview (t, a, b, c)         -> Mcallview (ft t, f a, g b, f c)
   | Mself id                       -> Mself (g id)
   (* operation *)
   | Moperations                    -> Moperations
@@ -2218,6 +2221,7 @@ let fold_term (f : 'a -> ('id mterm_gen) -> 'a) (accu : 'a) (term : 'id mterm_ge
   | Mtransfer tr                          -> fold_transfer_kind f accu tr
   (* entrypoint *)
   | Mentrypoint (_, _, s, r)              -> let tmp = f accu s in Option.map_dfl (f tmp) tmp r
+  | Mcallview (_, a, _, c)                -> f (f accu a) c
   | Mself _                               -> accu
   (* operation *)
   | Moperations                           -> accu
@@ -2660,6 +2664,11 @@ let fold_map_term
       | None -> (None, sa)
     in
     g (Mentrypoint (t, a, se, re)), ra
+
+  | Mcallview (t, a, b, c) ->
+    let ae, aa = f accu a in
+    let ce, ca = f aa c in
+    g (Mcallview (t, ae, b, ce)), ca
 
   | Mself id ->
     g (Mself id), accu