+++ src/symbol.ml
@@ -148,6 +148,7 @@ let string_of_token = function
   | COLONEQUAL           -> "COLONEQUAL"
   | COLONCOLON           -> "COLONCOLON"
   | COLON                -> "COLON"
+  | CALLVIEW             -> "CALLVIEW"
   | CALLED               -> "CALLED"
   | CALL                 -> "CALL"
   | BYTES s              -> Printf.sprintf "BYTES(%s)" s
@@ -320,6 +321,7 @@ let string_of_symbol = function
   | X (T T_COLONEQUAL) -> ":="
   | X (T T_COLONCOLON) -> "::"
   | X (T T_COLON) -> ":"
+  | X (T T_CALLVIEW) -> "callview"
   | X (T T_CALLED) -> "called"
   | X (T T_CALL) -> "call"
   | X (T T_BYTES) -> "bytes"